pyproject.toml
[project]
name = "whisper-transcription"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "librosa>=0.10.0",
    "numpy==1.26.4",
    "openai-whisper>=20231117",
    "pandas>=1.3.0",
    "pydub>=0.25.1",
    "soundfile>=0.12.1",
    "streamlit>=1.27.0",
    "torch==2.8.0",
    "torchaudio==2.8.0",
    "whisperx>=3.3.1",
]
